group commerce solution types under commerce app
