+++ packages/samples/headless-react/src/App.tsx
@@ -1,13 +1,16 @@
-import {BrowserRouter, NavLink, Routes, Route} from 'react-router-dom';
+import {BrowserRouter, NavLink, Route, Routes} from 'react-router-dom';
 import {AboutPage} from './pages/AboutPage';
 import {AnalyticsHook} from './pages/AnalyticsHook';
-import CommercePage from './pages/CommercePage';
 import {DependentFacetPage} from './pages/DependentFacetPage';
 import {ProductRecommendationsPage} from './pages/ProductRecommendationsPage';
 import {RecommendationPage} from './pages/RecommendationPage';
 import {SamlPage} from './pages/SamlPage';
 import {SearchPage, SearchPageProps} from './pages/SearchPage';
 import {StandaloneSearchBoxPage} from './pages/StandaloneSearchBoxPage';
+import {CommerceApp} from './pages/commerce/CommerceApp';
+import {ProductListingPage} from './pages/commerce/ProductListingPage';
+import {RecommendationsPage} from './pages/commerce/RecommendationsPage';
+import {SearchPage as CommerceSearchPage} from './pages/commerce/SearchPage';
 
 function App(props: SearchPageProps) {
   const activeNavLink: React.CSSProperties = {color: 'red'};
@@ -104,7 +107,11 @@ function App(props: SearchPageProps) {
             path="/product-recommendations"
             element={<ProductRecommendationsPage />}
           />
-          <Route path="/commerce" element={<CommercePage />} />
+          <Route path="/commerce" element={<CommerceApp />}>
+            <Route path="search" element={<CommerceSearchPage />} />
+            <Route path="product-listing" element={<ProductListingPage />} />
+            <Route path="recommendations" element={<RecommendationsPage />} />
+          </Route>
           <Route path="/search-page" element={<SearchPage {...props} />} />
           <Route path="/" element={<SearchPage {...props} />} />
         </Routes>
